refactor: move the TargetGroup package to discovery/
