+++ discovery/openstack/openstack.go
@@ -24,7 +24,7 @@ import (
 	"github.com/gophercloud/gophercloud/openstack"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/pkg/targetgroup"
+	"github.com/prometheus/prometheus/discovery/targetgroup"
 	config_util "github.com/prometheus/prometheus/util/config"
 	yaml_util "github.com/prometheus/prometheus/util/yaml"
 )
