+++ retrieval/helpers_test.go
@@ -14,8 +14,8 @@
 package retrieval
 
 import (
+	"github.com/prometheus/prometheus/discovery/targetgroup"
 	"github.com/prometheus/prometheus/pkg/labels"
-	"github.com/prometheus/prometheus/pkg/targetgroup"
 	"github.com/prometheus/prometheus/storage"
 )
