+++ retrieval/manager.go
@@ -20,7 +20,7 @@ import (
 	"github.com/go-kit/kit/log/level"
 
 	"github.com/prometheus/prometheus/config"
-	"github.com/prometheus/prometheus/pkg/targetgroup"
+	"github.com/prometheus/prometheus/discovery/targetgroup"
 	"github.com/prometheus/prometheus/storage"
 )
